상태관리와 정보조회하는 api함수 mypageForm으로 이동
