+++ src/components/member/mypage/MyPageForm.jsx
@@ -1,13 +1,43 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import styles from './MyPageForm.module.css';
 import Profile from './profile/Profile';
 import LikeList from './likeList/LikeList';
 import InfoEdit from './InfoEdit';
+import { getUserProfile } from 'utils/api/profileApi.js';
 
 function MyPageForm() {
     const [activeTab, setActiveTab] = useState('profile'); // 기본값: profile
 
     const [isEdit, setIsEdit] = useState(false);
+
+    const [infoData, setInfoData] = useState(
+        {
+            email:'', 
+            nickname:'', 
+            birth:'', 
+            phone:'', 
+            gender:'', 
+            height:'', 
+            weight:'', 
+        }
+    );
+
+    useEffect( () => {
+        const fetchProfile = async () => {
+            // 회원의 프로필 정보를 가져오는 api 함수
+            const result = await getUserProfile();
+            if (result.gender === 'F') {
+                result.gender = '여자';
+            } else if (result.gender === 'M') {
+                result.gender = '남자자';
+            } else {
+                // 정보가 잘못 저장되어 있으므로 따로 처리해줘야 함: TODO
+            }
+            setInfoData(result);
+        };
+
+        fetchProfile();
+    }, [isEdit]);
 
     const handleClickProfileTab = () => {
         if (!isEdit) {
@@ -43,7 +73,7 @@ function MyPageForm() {
                     {/* {activeTab === 'profile' ? <Profile /> : <LikeList />}
                     <InfoEdit /> */}
                     {
-                        isEdit === false ? (activeTab ==='profile' ? <Profile setIsEdit={setIsEdit} /> : <LikeList />) : <InfoEdit />
+                        isEdit === false ? (activeTab ==='profile' ? <Profile infoData={infoData} setIsEdit={setIsEdit} /> : <LikeList />) : <InfoEdit />
                     }
                 </div>
             </div>
